rename DurableOrchestrationClient to DurableClient
